[CastIt.Android] Clean the house part V
