+++ CastIt.Android/lib/main.dart
@@ -1,48 +1,28 @@
+import 'package:castit/domain/services/device_info_service.dart';
+import 'package:castit/domain/services/locale_service.dart';
+import 'package:castit/domain/services/logging_service.dart';
+import 'package:castit/domain/services/settings_service.dart';
+import 'package:castit/generated/l10n.dart';
+import 'package:castit/injection.dart';
+import 'package:castit/presentation/shared/extensions/app_theme_type_extensions.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 
-import 'bloc/intro/intro_bloc.dart';
-import 'bloc/main/main_bloc.dart';
-import 'bloc/play/play_bloc.dart';
-import 'bloc/played_file_options/played_file_options_bloc.dart';
-import 'bloc/playlist/playlist_bloc.dart';
-import 'bloc/playlist_rename/playlist_rename_bloc.dart';
-import 'bloc/playlists/playlists_bloc.dart';
-import 'bloc/server_ws/server_ws_bloc.dart';
-import 'bloc/settings/settings_bloc.dart';
-import 'generated/i18n.dart';
-import 'injection.dart';
-import 'services/logging_service.dart';
-import 'services/settings_service.dart';
-import 'telemetry.dart';
-import 'ui/pages/intro_page.dart';
-import 'ui/pages/main_page.dart';
+import 'application/bloc.dart';
+import 'presentation/intro/intro_page.dart';
+import 'presentation/main/main_page.dart';
 
 Future main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  initInjection();
-  await initTelemetry();
+  await initInjection();
   await SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]);
   runApp(MyApp());
 }
 
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  final GeneratedLocalizationsDelegate i18n = I18n.delegate;
-
-  @override
-  void initState() {
-    super.initState();
-    I18n.onLocaleChanged = _onLocaleChange;
-  }
-
+class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MultiBlocProvider(
@@ -51,7 +31,9 @@ class _MyAppState extends State<MyApp> {
           create: (ctx) {
             final logger = getIt<LoggingService>();
             final settings = getIt<SettingsService>();
-            return MainBloc(logger, settings)..add(MainEvent.init());
+            final deviceInfo = getIt<DeviceInfoService>();
+            final localeService = getIt<LocaleService>();
+            return MainBloc(logger, settings, deviceInfo, localeService)..add(MainEvent.init());
           },
         ),
         BlocProvider(
@@ -83,7 +65,7 @@ class _MyAppState extends State<MyApp> {
           create: (ctx) {
             final settings = getIt<SettingsService>();
             final serverWsBloc = ctx.read<ServerWsBloc>();
-            return SettingsBloc(settings, serverWsBloc);
+            return SettingsBloc(settings, serverWsBloc, ctx.read<MainBloc>());
           },
         ),
         BlocProvider(create: (ctx) {
@@ -105,13 +87,9 @@ class _MyAppState extends State<MyApp> {
 
   Widget _buildApp(MainState state) {
     final delegates = <LocalizationsDelegate>[
-      // A class which loads the translations from JSON files
-      i18n,
-      // Built-in localization of basic text for Material widgets
+      S.delegate,
       GlobalMaterialLocalizations.delegate,
-      // Built-in localization for text direction LTR/RTL
       GlobalWidgetsLocalizations.delegate,
-      // Built-in localization of basic text for Cupertino widgets
       GlobalCupertinoLocalizations.delegate,
     ];
     return state.map<Widget>(
@@ -119,23 +97,18 @@ class _MyAppState extends State<MyApp> {
         return const CircularProgressIndicator();
       },
       loaded: (s) {
+        final locale = Locale(s.language.code, s.language.countryCode);
+        final themeData = s.accentColor.getThemeData(s.theme);
         return MaterialApp(
           title: s.appTitle,
-          theme: s.theme,
+          theme: themeData,
           home: s.firstInstall ? IntroPage() : MainPage(),
+          //Without this, the lang won't be reloaded
+          locale: locale,
           localizationsDelegates: delegates,
-          supportedLocales: i18n.supportedLocales,
-          localeResolutionCallback: i18n.resolution(
-            fallback: i18n.supportedLocales.first,
-          ),
+          supportedLocales: S.delegate.supportedLocales,
         );
       },
     );
-  }
-
-  void _onLocaleChange(Locale locale) {
-    setState(() {
-      I18n.locale = locale;
-    });
   }
 }
