+++ CastIt.Android/lib/presentation/playlist/widgets/item_counter.dart
