+++ CastIt.Android/lib/presentation/shared/extensions/app_theme_type_extensions.dart
@@ -1,0 +1,103 @@
+import 'package:castit/domain/enums/enums.dart';
+import 'package:flutter/material.dart';
+
+extension AppThemeTypeExtensions on AppAccentColorType {
+  Color getAccentColor() {
+    switch (this) {
+      case AppAccentColorType.blue:
+        return Colors.blue;
+      case AppAccentColorType.green:
+        return Colors.green;
+      case AppAccentColorType.pink:
+        return Colors.pink;
+      case AppAccentColorType.brown:
+        return Colors.brown;
+      case AppAccentColorType.red:
+        return Colors.red;
+      case AppAccentColorType.cyan:
+        return Colors.cyan;
+      case AppAccentColorType.indigo:
+        return Colors.indigo;
+      case AppAccentColorType.purple:
+        return Colors.purple;
+      case AppAccentColorType.deepPurple:
+        return Colors.deepPurple;
+      case AppAccentColorType.grey:
+        return Colors.grey;
+      case AppAccentColorType.orange:
+        return Colors.orange;
+      case AppAccentColorType.yellow:
+        return Colors.yellow;
+      case AppAccentColorType.blueGrey:
+        return Colors.blueGrey;
+      case AppAccentColorType.teal:
+        return Colors.teal;
+      case AppAccentColorType.amber:
+        return Colors.amber;
+      default:
+        throw Exception('The provided accent color = $this is not valid ');
+    }
+  }
+
+  ThemeData getThemeData(AppThemeType theme) {
+    final color = getAccentColor();
+
+    switch (theme) {
+      case AppThemeType.dark:
+        return ThemeData(
+          visualDensity: VisualDensity.adaptivePlatformDensity,
+          brightness: Brightness.dark,
+          accentColor: color,
+          primaryColor: color,
+          primaryColorLight: color.withOpacity(0.5),
+          primaryColorDark: color,
+          primarySwatch: MaterialColor(color.value, getSwatch(color)),
+        );
+      case AppThemeType.light:
+        return ThemeData(
+          visualDensity: VisualDensity.adaptivePlatformDensity,
+          brightness: Brightness.light,
+          accentColor: color,
+          primaryColor: color,
+          primaryColorLight: color.withOpacity(0.5),
+          primaryColorDark: color,
+          primarySwatch: MaterialColor(color.value, getSwatch(color)),
+        );
+      default:
+        throw Exception('The provided theme  = $theme is not valid ');
+    }
+  }
+
+  Map<int, Color> getSwatch(Color color) {
+    final hslColor = HSLColor.fromColor(color);
+    final lightness = hslColor.lightness;
+
+    /// if [500] is the default color, there are at LEAST five
+    /// steps below [500]. (i.e. 400, 300, 200, 100, 50.) A
+    /// divisor of 5 would mean [50] is a lightness of 1.0 or
+    /// a color of #ffffff. A value of six would be near white
+    /// but not quite.
+    const lowDivisor = 6;
+
+    /// if [500] is the default color, there are at LEAST four
+    /// steps above [500]. A divisor of 4 would mean [900] is
+    /// a lightness of 0.0 or color of #000000
+    const highDivisor = 5;
+
+    final lowStep = (1.0 - lightness) / lowDivisor;
+    final highStep = lightness / highDivisor;
+
+    return {
+      50: (hslColor.withLightness(lightness + (lowStep * 5))).toColor(),
+      100: (hslColor.withLightness(lightness + (lowStep * 4))).toColor(),
+      200: (hslColor.withLightness(lightness + (lowStep * 3))).toColor(),
+      300: (hslColor.withLightness(lightness + (lowStep * 2))).toColor(),
+      400: (hslColor.withLightness(lightness + lowStep)).toColor(),
+      500: (hslColor.withLightness(lightness)).toColor(),
+      600: (hslColor.withLightness(lightness - highStep)).toColor(),
+      700: (hslColor.withLightness(lightness - (highStep * 2))).toColor(),
+      800: (hslColor.withLightness(lightness - (highStep * 3))).toColor(),
+      900: (hslColor.withLightness(lightness - (highStep * 4))).toColor(),
+    };
+  }
+}
