+++ CastIt.Android/pubspec.yaml
@@ -23,6 +23,7 @@ environment:
 dependencies:
   cached_network_image: ^3.0.0
   collection: ^1.15.0
+  device_info_plus: ^2.0.1
   flutter:
     sdk: flutter
   flutter_bloc: ^7.0.1
